add van model types
